+++ reapy/core/window/__init__.py
@@ -1,0 +1,2 @@
+from .window import Window
+from .midi_editor import MIDIEditor