+++ reapy/core/window/window.py
@@ -1,0 +1,15 @@
+import reapy
+from reapy.core import ReapyObject
+import reapy.reascript_api as RPR
+
+
+class Window(ReapyObject):
+
+    _class_name = "Window"
+
+    def __init__(self, id):
+        self.id = id
+
+    @property
+    def _args(self):
+        return (self.id,)